vp8 enc: multi-thread the auto-filter statistics
The LFStats SSIM accumulators become fixed-point integers (32 fractional
bits, rounded once per macroblock and filter level): integer addition is
exact in any order, so the wavefront workers can accumulate their rows
privately and merge in any order while staying bit-exact with the
single-threaded loop. The trial filtering itself is per-macroblock
independent (it only filters the inner sub-block edges of the private
reconstruction buffers), so no extra synchronization is needed.

The single-threaded output is unchanged in practice: the 2^-32
quantization is noise next to the 1e-5 relative decision threshold
(byte-identical files on all tested images and settings). The worst-case
sum (172 SSIM windows per macroblock, 16383x16383 image) fits int64_t
with an order of magnitude to spare.

Auto-filter was the last mode falling back to the single-threaded token
loop, and it benefits the most from the wavefront since the level search
dominates the encoding cost: 12k photo q75 -af on an M4 Pro goes from
63.9s to 9.8s with -mt (6.5x), byte-identical to the single-threaded
file. ThreadSanitizer-clean, deterministic across runs, validated over
q5-q95 x m3-m6 x three content types plus size-search/multi-pass paths,
on arm64 and x86-64.

+++ src/enc/vp8i_enc.h
@@ -135,7 +135,10 @@ typedef proba_t StatsArray[NUM_CTX][NUM_PROBAS];
 typedef uint16_t CostArray[NUM_CTX][MAX_VARIABLE_LEVEL + 1];
 typedef const uint16_t* (*CostArrayPtr)[NUM_CTX];  // for easy casting
 typedef const uint16_t* CostArrayMap[16][NUM_CTX];
-typedef double LFStats[NUM_MB_SEGMENTS][MAX_LF_LEVELS];  // filter stats
+// Filter stats: cumulated SSIM scores per segment and filter level, in
+// LF_STATS_FIX fixed-point. Integer accumulation is exact in any order,
+// which keeps the multi-threaded loop bit-exact with the single-threaded one.
+typedef int64_t LFStats[NUM_MB_SEGMENTS][MAX_LF_LEVELS];
 
 typedef struct VP8Encoder VP8Encoder;
